+++ config.php
@@ -3,13 +3,7 @@
 
 /* Function-based configuration
 ================== */
-
-// ----------------- //
-$dbusers = "db/users.json"; //It is *RECOMMENDED* that this file is renamed and/or moved, for maximum security -- it is the file in which user information is stored
-// ----------------- //
-
 $triggerclass = "slum"; //Class that makes elements editable by Slum
-$triggertitle = "slumtitle"; //Change the attributename for defining titles (e.g. slumtitle="First column", where slumtitle is triggertitle)
 
 /* Text-area
 ================== */
@@ -23,9 +17,12 @@ $wysiwygbuttonconfig = "'bold','italic','underline','strikethrough','left','cent
 $showpath = "false"; //Show path to the page underneath its name
 
 //Less important configurations
+$triggertitle = "slumtitle"; //Change the attributename for defining titles (e.g. slumtitle="First column", where slumtitle is triggertitle)
 $triggertextarea = "textarea"; //Textarea defining attribute for HTML elements (e.g. textarea="plain", where textarea is the attribute)
 $rootpath = $_SERVER['DOCUMENT_ROOT'] . "/"; //Makes all page URLS from webroot instead of relative to the Slum directory
 $dbpages = "db/pages.json"; //Path to the database file containing information regarding pages
+$dbusers = "db/users.php"; //Path to database file containing information regarding users -- is protected (notice .php extension, if you change it's location or name)
+
 
 /* Promotional-based configuration
 (Change this if you don't want it to say Slum CMS all over.)